Move util into internal package
