+++ pkg/imgpkg/internal/util/prefixed_logger.go
