+++ pkg/imgpkg/internal/util/progress_logger.go
