+++ pkg/imgpkg/internal/util/tag.go
@@ -5,6 +5,7 @@ package util
 
 import (
 	"fmt"
+
 	regname "github.com/google/go-containerregistry/pkg/name"
 	regv1 "github.com/google/go-containerregistry/pkg/v1"
 )
